+++ src/main/java/com/proiect/ProiectIsAeroport/companie/Zbor_Sezonier.java
@@ -14,7 +14,7 @@ public class Zbor_Sezonier extends Zbor_Regulat {
         super();
     }
 
-    public Zbor_Sezonier(String cod_cursa, Tip_Zbor tip_zbor, String oras_destinatie, String oras_plecare, int tarifeBusiness, int tarifeClasa1, int tarifeEconomie, Model model, int locuriBusiness, int locuriClasa1, int locuriEconomie, boolean esteTurRetur, int discount, String zi, String ora, String inceput, String sfarsit) {
+    public Zbor_Sezonier(String cod_cursa, Tip_Zbor tip_zbor, String oras_destinatie, String oras_plecare, double tarifeBusiness, double tarifeClasa1, double tarifeEconomie, Model model, int locuriBusiness, int locuriClasa1, int locuriEconomie, boolean esteTurRetur, int discount, String zi, String ora, String inceput, String sfarsit) {
         super(cod_cursa, tip_zbor, oras_destinatie, oras_plecare, tarifeBusiness, tarifeClasa1, tarifeEconomie, model, locuriBusiness, locuriClasa1, locuriEconomie, esteTurRetur, discount, zi, ora);
         this.inceput = inceput;
         this.sfarsit = sfarsit;
